Many improvements and fixes
* Separate python binding: an example/ directory can be used now to test and debug the program without having to load python.
* Better error handling
* Fixed threading issues: data races with `end` and in the pthread_cond_t wait loop. Also initializing the mutex in cross_correlation.c.
* Fixed memory leaks: `data1` and `data2` were not freed, amongst other things. This also closes #5.

+++ src/global.h
@@ -26,7 +26,7 @@ struct thread_data {
     size_t total_len;
     size_t len;
     const size_t *intervals;
-    const int n_intervals;
+    const size_t n_intervals;
     pthread_mutex_t *mutex;
     pthread_cond_t *done;
     int *end;
